backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

langchain
langchain-community
langchain-text-splitters

langchain-huggingface
sentence-transformers

langchain-pinecone
pinecone[grpc]

pypdf
docx2txt
